hbrmt: change pixel argument types to uintptr_t
The reasoning for this is:
- We want it to be unsigned because we don't expect a negative value to
  be passed.  We perform unsigned shifts on the value in some cases.
- We want it to be register sized so that we do not have to extend it on
  x86-64.

This also removes some const qualifiers from the argument in some cases.

+++ lib/upipe-hbrmt/sdidec.c
@@ -1,7 +1,7 @@
 #include <inttypes.h>
 #include "sdidec.h"
 
-void upipe_sdi_unpack_c(const uint8_t *src, uint16_t *y, int64_t pixels)
+void upipe_sdi_unpack_c(const uint8_t *src, uint16_t *y, uintptr_t pixels)
 {
     for (int i = 0; i < pixels; i += 4) {
         uint8_t a = *src++;
@@ -16,7 +16,7 @@ void upipe_sdi_unpack_c(const uint8_t *src, uint16_t *y, int64_t pixels)
     }
 }
 
-void upipe_sdi_v210_unpack_c(const uint8_t *src, uint32_t *dst, int64_t pixels)
+void upipe_sdi_v210_unpack_c(const uint8_t *src, uint32_t *dst, uintptr_t pixels)
 {
     for (int i = 0; i < pixels; i += 3 * 4) {
         uint16_t a, b, c;
@@ -43,7 +43,7 @@ void upipe_sdi_v210_unpack_c(const uint8_t *src, uint32_t *dst, int64_t pixels)
     }
 }
 
-void upipe_sdi_to_planar_8_c(const uint8_t *src, uint8_t *y, uint8_t *u, uint8_t *v, int64_t pixels)
+void upipe_sdi_to_planar_8_c(const uint8_t *src, uint8_t *y, uint8_t *u, uint8_t *v, uintptr_t pixels)
 {
     for (int i = 0; i < pixels; i += 4) {
         uint8_t a = *src++; // UUUUUUUU
